separated BidSelectionHeader to own file
