+++ src/components/BidSelectionHeader.css
@@ -1,0 +1,62 @@
+.bid-selection-header-wrapper {
+  min-height: 60px;
+}
+
+.bid-selection-header {
+  height: 60px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: absolute;
+  width: 100%;
+  top: 0;
+}
+
+.bid-selection-header__back-button {
+  margin-left: 10px;
+  padding: 10px;
+}
+
+.bid-selection-header__title {
+}
+
+.bid-selection-header__sort-button {
+  margin-right: 10px;
+  padding: 10px;
+}
+
+.sort-options {
+  position: relative;
+  background: #ffffff;
+  text-align: center;
+  box-shadow: 0 2px 10px 0 #c4c8fa;
+  height: 0;
+  transition: all 0.5s ease-in-out;
+  overflow-y: hidden;
+  z-index: 100;
+}
+
+.sort-options--expand {
+  height: 180px;
+}
+
+.sort-options__close-button {
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 10px;
+  margin: 10px 13px 0 0;
+}
+
+.sort-options__header {
+  padding: 20px 0 15px 0;
+}
+
+.sort-options__list {
+  /* padding-bottom: 15px; */
+}
+
+.sort-options__list > li {
+  margin: 0 0 10px 0;
+  font-size: 13px;
+}
